requirements.txt
torch==2.6.0
torchvision==0.21.0
numpy==2.0.2
pandas==2.3.1
pillow==10.4.0
opencv-python==4.12.0.88
requests==2.32.4
flask==3.1.2
gradio==4.44.1
matplotlib==3.9.4
diffusers
transformers
openai==1.95.1
beautifulsoup4==4.13.4
pydantic==2.10.6
jinja2==3.1.6
python-dotenv==1.1.1
tqdm==4.67.1
ddgs==9.5.4